WIP: partially implement caching!
Progress needs to have a boolean for enabling adding to
history or not for when we go back in History. Currently
History does not use Progress because it avoids adding to
history when you go back/forward.

Caching is pretty much done being implemented otherwise!
Could use more documentation, refactory, etc.