WIP: fix viewport bug
This is buggy because I should have a different event for signifying
that we're done with displaying the progress and now we're going to
switch to a new mode... I guess? I don't know... maybe just a flag
in the state... I'm not sure how to tackle this. The problem is right
now the modeTransition function gets executed every update to the state
via NewState event which gets updated to show progress in downloading bytes
until the final state which is the new page (also sent via NewState event).