Fix gopher URIs (gophertype)
In RFC 4266 the path/resource in a Gopher URI has a
"gopheritem" prefix in the format /1/foo/bar (`1`
being the gopheritem). This gopheritem prefix allows
clients to decipher the content the URI points to,
a job usually fulfilled by a menu's item type
(RFC 1436).

Implement a function which removes the "gopheritem"
prefix in the resource part of the URI, but still use
the "gopheritem" prefix to determine the type of
content the URI points to.

Implement a function which gets the "gopheritem"
(a single character) from a selector/resource.

Use these new tools for `GotoMode` and for
the starting location provided when executing
Waffle.

Remove `itemTypeToRenderMode` and `guessMode`
in favor for the new tools used with the new
`selectorToRenderMode` function which will
give a `RenderMode` for the supplied URI.